backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.8.2
pydantic[email]==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.2
anyio==4.4.0
pydantic-settings